Now possible to use parent's method in subclass
This make the program easier to read and maintain.
When you want to do a RPC using wamp: you don't have anymore to pass a message with pyqtsignal to the Mainwindows class, set up a function to deal with its contents and send the result using an other pyqtsignal. You can directly call the parent "call" function in the subclass, same for all the other functions.